backend/requirements.txt
# SomniProperty Backend - Python Dependencies

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
PyJWT==2.8.0

# MQTT
paho-mqtt==1.6.1

# Redis for shared caching and WebSocket token storage
redis==5.0.1

# HTTP client for AI providers
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.5

# Web scraping for vendor pricing
beautifulsoup4==4.12.2
lxml==4.9.3

# AI Provider SDKs
# openai==1.12.0
anthropic==0.18.1  # For Claude log analysis in HA Instance service

# Encryption for HA Instance API tokens
cryptography==42.0.2

# Stripe payments
stripe==7.8.0

# MinIO object storage
minio==7.2.0

# Image processing for thumbnails
Pillow==10.1.0

# MIME type detection
python-magic==0.4.27

# PDF generation
weasyprint==60.1
pydyf<0.11.0  # WeasyPrint 60.1 requires pydyf < 0.11.0
qrcode[pil]==7.4.2

# Utilities
python-dotenv==1.0.0
psutil==5.9.6

# SMS and Notifications
twilio==8.11.0

# Template engine for K8s manifests
jinja2==3.1.2

# GitOps integration
GitPython==3.1.40

# YAML parsing for service catalog
PyYAML==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiosqlite==0.19.0
sendgrid==6.11.0
